Add util for fixerState and separate subcomponents
